update: 更新文件 net_http_probe.cpp

+++ services/netconnmanager/src/net_http_probe.cpp
@@ -720,7 +720,7 @@ int32_t NetHttpProbe::CheckSuccessRespCode(int32_t respCode)
         return result;
     }
     std::string requestId = GetHeaderField(xReqId_);
-    if (requestId.empty() || xReqIdLen_ != static_cast<int32_t>(requestId.length())) {
+    if (requestId.empty() || xReqIdLen_ != static_cast<int8_t>(requestId.length())) {
         NETMGR_LOG_I("http return 204, but request id error and unreachable!");
         result = FAIL_CODE;
     }
